moved all helper functions
