+++ dfirtrack_main/tests/system_importer/test_system_importer_file_csv_check_config_attributes.py
@@ -5,302 +5,16 @@ from dfirtrack.settings import BASE_DIR
 from dfirtrack_config.models import SystemImporterFileCsvConfigModel
 from dfirtrack_main.importer.file.csv import system_cron
 from dfirtrack_main.models import Analysisstatus, Case, Company, Dnsname, Domain, Location, Os, Reason, Recommendation, Serviceprovider, Systemstatus, Systemtype, Tag, Tagcolor
+from dfirtrack_main.tests.system_importer.config_functions import set_config_check_config_attributes_column_fields_numeric_values
+from dfirtrack_main.tests.system_importer.config_functions import set_config_check_config_attributes_column_choice_vs_default_single_error
+from dfirtrack_main.tests.system_importer.config_functions import set_config_check_config_attributes_column_choice_vs_default_multiple_errors_1
+from dfirtrack_main.tests.system_importer.config_functions import set_config_check_config_attributes_column_choice_vs_default_multiple_errors_2
+from dfirtrack_main.tests.system_importer.config_functions import set_config_check_config_attributes_tagfree_choices
+from dfirtrack_main.tests.system_importer.config_functions import set_config_check_config_attributes_column_fields_equal_values
+from dfirtrack_main.tests.system_importer.config_functions import set_config_check_config_attributes_remove_choices
 import os
 import urllib.parse
 
-
-def set_config_choices_true(system_importer_file_csv_config_model):
-    """ set choices to true for all columns """
-
-    # set config values
-    system_importer_file_csv_config_model.csv_choice_system = True
-    system_importer_file_csv_config_model.csv_choice_ip = True
-    system_importer_file_csv_config_model.csv_choice_dnsname = True
-    system_importer_file_csv_config_model.csv_choice_domain = True
-    system_importer_file_csv_config_model.csv_choice_location = True
-    system_importer_file_csv_config_model.csv_choice_os = True
-    system_importer_file_csv_config_model.csv_choice_reason = True
-    system_importer_file_csv_config_model.csv_choice_recommendation = True
-    system_importer_file_csv_config_model.csv_choice_serviceprovider = True
-    system_importer_file_csv_config_model.csv_choice_systemtype = True
-    system_importer_file_csv_config_model.csv_choice_case = True
-    system_importer_file_csv_config_model.csv_choice_company = True
-    system_importer_file_csv_config_model.csv_choice_tag = True
-
-    # return config to config function
-    return system_importer_file_csv_config_model
-
-def set_config_column_fields_numeric_values():
-    """ set numeric values for columns out of range """
-
-    # get config
-    system_importer_file_csv_config_model = SystemImporterFileCsvConfigModel.objects.get(system_importer_file_csv_config_name='SystemImporterFileCsvConfig')
-
-    # set choices to true for all columns
-    system_importer_file_csv_config_model = set_config_choices_true(system_importer_file_csv_config_model)
-
-    # set config values
-    system_importer_file_csv_config_model.csv_column_system = 100
-    system_importer_file_csv_config_model.csv_column_ip = 101
-    system_importer_file_csv_config_model.csv_column_dnsname = 102
-    system_importer_file_csv_config_model.csv_column_domain = 103
-    system_importer_file_csv_config_model.csv_column_location = 104
-    system_importer_file_csv_config_model.csv_column_os = 105
-    system_importer_file_csv_config_model.csv_column_reason = 106
-    system_importer_file_csv_config_model.csv_column_recommendation = 107
-    system_importer_file_csv_config_model.csv_column_serviceprovider = 108
-    system_importer_file_csv_config_model.csv_column_systemtype = 109
-    system_importer_file_csv_config_model.csv_column_case = 110
-    system_importer_file_csv_config_model.csv_column_company = 111
-    system_importer_file_csv_config_model.csv_column_tag = 112
-
-    # save config
-    system_importer_file_csv_config_model.save()
-
-    # return to column fields numeric values test function
-    return
-
-def set_config_column_choice_vs_default_single_error():
-    """ set column, choice and default single error """
-
-    # get config
-    system_importer_file_csv_config_model = SystemImporterFileCsvConfigModel.objects.get(system_importer_file_csv_config_name='SystemImporterFileCsvConfig')
-
-    # set config values
-    system_importer_file_csv_config_model.csv_column_system = 1
-    system_importer_file_csv_config_model.csv_choice_ip = False
-    system_importer_file_csv_config_model.csv_column_ip = 2
-
-    # save config
-    system_importer_file_csv_config_model.save()
-
-    # return to column choice vs default test function
-    return
-
-def set_config_column_choice_vs_default_multiple_errors_1():
-    """ set column, choice and default randomly faulty """
-
-    # get objects
-    case_1 = Case.objects.get(case_name='case_1')
-    company_1 = Company.objects.get(company_name='company_1')
-    dnsname_1 = Dnsname.objects.get(dnsname_name='dnsname_1')
-    domain_1 = Domain.objects.get(domain_name='domain_1')
-    location_1 = Location.objects.get(location_name='location_1')
-    reason_1 = Reason.objects.get(reason_name='reason_1')
-    recommendation_1 = Recommendation.objects.get(recommendation_name='recommendation_1')
-    serviceprovider_1 = Serviceprovider.objects.get(serviceprovider_name='serviceprovider_1')
-    systemtype_1 = Systemtype.objects.get(systemtype_name='systemtype_1')
-    os_1 = Os.objects.get(os_name='os_1')
-    tag_1 = Tag.objects.get(tag_name='tag_1')
-
-    # get config
-    system_importer_file_csv_config_model = SystemImporterFileCsvConfigModel.objects.get(system_importer_file_csv_config_name='SystemImporterFileCsvConfig')
-
-    # set config values
-    system_importer_file_csv_config_model.csv_column_system = 1
-    # CSV not chosen and CSV column filled out / +1
-    system_importer_file_csv_config_model.csv_choice_ip = False
-    system_importer_file_csv_config_model.csv_column_ip = 2
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_dnsname = False
-    system_importer_file_csv_config_model.csv_column_dnsname = 3
-    system_importer_file_csv_config_model.csv_default_dnsname = dnsname_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_domain = False
-    system_importer_file_csv_config_model.csv_column_domain = 4
-    system_importer_file_csv_config_model.csv_default_domain = domain_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_location = False
-    system_importer_file_csv_config_model.csv_column_location = 5
-    system_importer_file_csv_config_model.csv_default_location = location_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_os = False
-    system_importer_file_csv_config_model.csv_column_os = 6
-    system_importer_file_csv_config_model.csv_default_os = os_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_reason = False
-    system_importer_file_csv_config_model.csv_column_reason = 7
-    system_importer_file_csv_config_model.csv_default_reason = reason_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_recommendation = False
-    system_importer_file_csv_config_model.csv_column_recommendation = 8
-    system_importer_file_csv_config_model.csv_default_recommendation = recommendation_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_serviceprovider = False
-    system_importer_file_csv_config_model.csv_column_serviceprovider = 9
-    system_importer_file_csv_config_model.csv_default_serviceprovider = serviceprovider_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_systemtype = False
-    system_importer_file_csv_config_model.csv_column_systemtype = 10
-    system_importer_file_csv_config_model.csv_default_systemtype = systemtype_1
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_case = False
-    system_importer_file_csv_config_model.csv_column_case = 11
-    system_importer_file_csv_config_model.csv_default_case.add(case_1)
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_company = False
-    system_importer_file_csv_config_model.csv_column_company = 12
-    system_importer_file_csv_config_model.csv_default_company.add(company_1)
-    # CSV not chosen and CSV column filled out / CSV column filled out and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_tag = False
-    system_importer_file_csv_config_model.csv_column_tag = 13
-    system_importer_file_csv_config_model.csv_default_tag.add(tag_1)
-
-    # save config
-    system_importer_file_csv_config_model.save()
-
-    # return to column choice vs default test function
-    return
-
-def set_config_column_choice_vs_default_multiple_errors_2():
-    """ set column, choice and default randomly faulty """
-
-    # get objects
-    case_1 = Case.objects.get(case_name='case_1')
-    company_1 = Company.objects.get(company_name='company_1')
-    dnsname_1 = Dnsname.objects.get(dnsname_name='dnsname_1')
-    domain_1 = Domain.objects.get(domain_name='domain_1')
-    location_1 = Location.objects.get(location_name='location_1')
-    reason_1 = Reason.objects.get(reason_name='reason_1')
-    recommendation_1 = Recommendation.objects.get(recommendation_name='recommendation_1')
-    serviceprovider_1 = Serviceprovider.objects.get(serviceprovider_name='serviceprovider_1')
-    systemtype_1 = Systemtype.objects.get(systemtype_name='systemtype_1')
-    os_1 = Os.objects.get(os_name='os_1')
-    tag_1 = Tag.objects.get(tag_name='tag_1')
-
-    # get config
-    system_importer_file_csv_config_model = SystemImporterFileCsvConfigModel.objects.get(system_importer_file_csv_config_name='SystemImporterFileCsvConfig')
-
-    # set config values
-    system_importer_file_csv_config_model.csv_column_system = 1
-    # CSV chosen and no CSV column filled out / +1
-    system_importer_file_csv_config_model.csv_choice_ip = True
-    system_importer_file_csv_config_model.csv_column_ip = None
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_dnsname = True
-    system_importer_file_csv_config_model.csv_column_dnsname = None
-    system_importer_file_csv_config_model.csv_default_dnsname = dnsname_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_domain = True
-    system_importer_file_csv_config_model.csv_column_domain = None
-    system_importer_file_csv_config_model.csv_default_domain = domain_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_location = True
-    system_importer_file_csv_config_model.csv_column_location = None
-    system_importer_file_csv_config_model.csv_default_location = location_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_os = True
-    system_importer_file_csv_config_model.csv_column_os = None
-    system_importer_file_csv_config_model.csv_default_os = os_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_reason = True
-    system_importer_file_csv_config_model.csv_column_reason = None
-    system_importer_file_csv_config_model.csv_default_reason = reason_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_recommendation = True
-    system_importer_file_csv_config_model.csv_column_recommendation = None
-    system_importer_file_csv_config_model.csv_default_recommendation = recommendation_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_serviceprovider = True
-    system_importer_file_csv_config_model.csv_column_serviceprovider = None
-    system_importer_file_csv_config_model.csv_default_serviceprovider = serviceprovider_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_systemtype = True
-    system_importer_file_csv_config_model.csv_column_systemtype = None
-    system_importer_file_csv_config_model.csv_default_systemtype = systemtype_1
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_case = True
-    system_importer_file_csv_config_model.csv_column_case = None
-    system_importer_file_csv_config_model.csv_default_case.add(case_1)
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_company = True
-    system_importer_file_csv_config_model.csv_column_company = None
-    system_importer_file_csv_config_model.csv_default_company.add(company_1)
-    # CSV chosen and no CSV column filled out / CSV chosen and DB chosen / +2
-    system_importer_file_csv_config_model.csv_choice_tag = True
-    system_importer_file_csv_config_model.csv_column_tag = None
-    system_importer_file_csv_config_model.csv_default_tag.add(tag_1)
-
-    # save config
-    system_importer_file_csv_config_model.save()
-
-    # return to column choice vs default test function
-    return
-
-def set_config_tagfree_choices():
-    """ set tagfree statuses without tag choice """
-
-    # get config
-    system_importer_file_csv_config_model = SystemImporterFileCsvConfigModel.objects.get(system_importer_file_csv_config_name='SystemImporterFileCsvConfig')
-
-    # set config values
-    system_importer_file_csv_config_model.csv_choice_tag = False
-    system_importer_file_csv_config_model.csv_choice_tagfree_systemstatus = True
-    system_importer_file_csv_config_model.csv_choice_tagfree_analysisstatus = True
-
-    # save config
-    system_importer_file_csv_config_model.save()
-
-    # return to tagfree choices test function
-    return
-
-def set_config_column_fields_equal_values():
-    """ set numeric values for column fields equal """
-
-    # get config
-    system_importer_file_csv_config_model = SystemImporterFileCsvConfigModel.objects.get(system_importer_file_csv_config_name='SystemImporterFileCsvConfig')
-
-    # set choices to true for all columns
-    system_importer_file_csv_config_model = set_config_choices_true(system_importer_file_csv_config_model)
-
-    # set config values
-    system_importer_file_csv_config_model.csv_column_system = 1
-    system_importer_file_csv_config_model.csv_column_ip = 1
-    system_importer_file_csv_config_model.csv_column_dnsname = 1
-    system_importer_file_csv_config_model.csv_column_domain = 1
-    system_importer_file_csv_config_model.csv_column_location = 1
-    system_importer_file_csv_config_model.csv_column_os = 1
-    system_importer_file_csv_config_model.csv_column_reason = 1
-    system_importer_file_csv_config_model.csv_column_recommendation = 1
-    system_importer_file_csv_config_model.csv_column_serviceprovider = 1
-    system_importer_file_csv_config_model.csv_column_systemtype = 1
-    system_importer_file_csv_config_model.csv_column_case = 1
-    system_importer_file_csv_config_model.csv_column_company = 1
-    system_importer_file_csv_config_model.csv_column_tag = 1
-
-    # save config
-    system_importer_file_csv_config_model.save()
-
-    # return to column fields different values test function
-    return
-
-def set_config_remove_choices():
-    """ set remove choices in combination with skipping systems """
-
-    # get config
-    system_importer_file_csv_config_model = SystemImporterFileCsvConfigModel.objects.get(system_importer_file_csv_config_name='SystemImporterFileCsvConfig')
-
-    # set config values
-    system_importer_file_csv_config_model.csv_skip_existing_system = True
-    system_importer_file_csv_config_model.csv_remove_systemstatus = True
-    system_importer_file_csv_config_model.csv_remove_analysisstatus = True
-    system_importer_file_csv_config_model.csv_remove_ip = True
-    system_importer_file_csv_config_model.csv_remove_dnsname = True
-    system_importer_file_csv_config_model.csv_remove_domain = True
-    system_importer_file_csv_config_model.csv_remove_location = True
-    system_importer_file_csv_config_model.csv_remove_os = True
-    system_importer_file_csv_config_model.csv_remove_reason = True
-    system_importer_file_csv_config_model.csv_remove_recommendation = True
-    system_importer_file_csv_config_model.csv_remove_serviceprovider = True
-    system_importer_file_csv_config_model.csv_remove_systemtype = True
-    system_importer_file_csv_config_model.csv_remove_case = True
-    system_importer_file_csv_config_model.csv_remove_company = True
-
-    # save config
-    system_importer_file_csv_config_model.save()
-
-    # return to remove choices test function
-    return
 
 class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
     """ system importer file CSV view tests """
@@ -441,7 +155,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_numeric_values()
+        set_config_check_config_attributes_column_fields_numeric_values()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -481,7 +195,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         """ test importer view """
 
         # change config
-        set_config_column_fields_numeric_values()
+        set_config_check_config_attributes_column_fields_numeric_values()
         # execute cron job / scheduled task
         system_cron()
         # login testuser
@@ -512,7 +226,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_numeric_values()
+        set_config_check_config_attributes_column_fields_numeric_values()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -554,7 +268,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_numeric_values()
+        set_config_check_config_attributes_column_fields_numeric_values()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -596,7 +310,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_numeric_values()
+        set_config_check_config_attributes_column_fields_numeric_values()
         # open upload file
         systemcsv = open(os.path.join(BASE_DIR, 'dfirtrack_main/tests/system_importer/system_importer_file_csv_files/system_importer_file_csv_testfile_01_minimal_double_quotation.csv'), 'r')
         # create post data
@@ -648,7 +362,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_single_error()
+        set_config_check_config_attributes_column_choice_vs_default_single_error()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -664,7 +378,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         """ test importer view """
 
         # change config
-        set_config_column_choice_vs_default_single_error()
+        set_config_check_config_attributes_column_choice_vs_default_single_error()
         # execute cron job / scheduled task
         system_cron()
         # login testuser
@@ -695,7 +409,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_single_error()
+        set_config_check_config_attributes_column_choice_vs_default_single_error()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -713,7 +427,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_single_error()
+        set_config_check_config_attributes_column_choice_vs_default_single_error()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -731,7 +445,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_single_error()
+        set_config_check_config_attributes_column_choice_vs_default_single_error()
         # open upload file
         systemcsv = open(os.path.join(BASE_DIR, 'dfirtrack_main/tests/system_importer/system_importer_file_csv_files/system_importer_file_csv_testfile_01_minimal_double_quotation.csv'), 'r')
         # create post data
@@ -759,7 +473,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_1()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_1()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -775,7 +489,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         """ test importer view """
 
         # change config
-        set_config_column_choice_vs_default_multiple_errors_1()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_1()
         # execute cron job / scheduled task
         system_cron()
         # login testuser
@@ -806,7 +520,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_1()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_1()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -824,7 +538,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_1()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_1()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -842,7 +556,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_1()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_1()
         # open upload file
         systemcsv = open(os.path.join(BASE_DIR, 'dfirtrack_main/tests/system_importer/system_importer_file_csv_files/system_importer_file_csv_testfile_01_minimal_double_quotation.csv'), 'r')
         # create post data
@@ -870,7 +584,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_2()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_2()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -886,7 +600,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         """ test importer view """
 
         # change config
-        set_config_column_choice_vs_default_multiple_errors_2()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_2()
         # execute cron job / scheduled task
         system_cron()
         # login testuser
@@ -917,7 +631,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_2()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_2()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -935,7 +649,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_2()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_2()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -953,7 +667,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_choice_vs_default_multiple_errors_2()
+        set_config_check_config_attributes_column_choice_vs_default_multiple_errors_2()
         # open upload file
         systemcsv = open(os.path.join(BASE_DIR, 'dfirtrack_main/tests/system_importer/system_importer_file_csv_files/system_importer_file_csv_testfile_01_minimal_double_quotation.csv'), 'r')
         # create post data
@@ -1692,7 +1406,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_tagfree_choices()
+        set_config_check_config_attributes_tagfree_choices()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1710,7 +1424,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         """ test importer view """
 
         # change config
-        set_config_tagfree_choices()
+        set_config_check_config_attributes_tagfree_choices()
         # execute cron job / scheduled task
         system_cron()
         # login testuser
@@ -1741,7 +1455,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_tagfree_choices()
+        set_config_check_config_attributes_tagfree_choices()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1761,7 +1475,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_tagfree_choices()
+        set_config_check_config_attributes_tagfree_choices()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1781,7 +1495,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_tagfree_choices()
+        set_config_check_config_attributes_tagfree_choices()
         # open upload file
         systemcsv = open(os.path.join(BASE_DIR, 'dfirtrack_main/tests/system_importer/system_importer_file_csv_files/system_importer_file_csv_testfile_01_minimal_double_quotation.csv'), 'r')
         # create post data
@@ -1811,7 +1525,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_equal_values()
+        set_config_check_config_attributes_column_fields_equal_values()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1827,7 +1541,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         """ test importer view """
 
         # change config
-        set_config_column_fields_equal_values()
+        set_config_check_config_attributes_column_fields_equal_values()
         # execute cron job / scheduled task
         system_cron()
         # login testuser
@@ -1858,7 +1572,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_equal_values()
+        set_config_check_config_attributes_column_fields_equal_values()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1876,7 +1590,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_equal_values()
+        set_config_check_config_attributes_column_fields_equal_values()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1894,7 +1608,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_column_fields_equal_values()
+        set_config_check_config_attributes_column_fields_equal_values()
         # open upload file
         systemcsv = open(os.path.join(BASE_DIR, 'dfirtrack_main/tests/system_importer/system_importer_file_csv_files/system_importer_file_csv_testfile_01_minimal_double_quotation.csv'), 'r')
         # create post data
@@ -1922,7 +1636,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_remove_choices()
+        set_config_check_config_attributes_remove_choices()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1938,7 +1652,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         """ test importer view """
 
         # change config
-        set_config_remove_choices()
+        set_config_check_config_attributes_remove_choices()
         # execute cron job / scheduled task
         system_cron()
         # login testuser
@@ -1969,7 +1683,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_remove_choices()
+        set_config_check_config_attributes_remove_choices()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -1987,7 +1701,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_remove_choices()
+        set_config_check_config_attributes_remove_choices()
         # create url
         destination = urllib.parse.quote('/system/', safe='/')
         # get response
@@ -2005,7 +1719,7 @@ class SystemImporterFileCsvCheckConfigAttributesViewTestCase(TestCase):
         # login testuser
         self.client.login(username='testuser_system_importer_file_csv_check_config_attributes', password='VgnTh4qkYZH61F5YArC7')
         # change config
-        set_config_remove_choices()
+        set_config_check_config_attributes_remove_choices()
         # open upload file
         systemcsv = open(os.path.join(BASE_DIR, 'dfirtrack_main/tests/system_importer/system_importer_file_csv_files/system_importer_file_csv_testfile_01_minimal_double_quotation.csv'), 'r')
         # create post data
